+++ internal/app/handlers/clicks/clicks_handler_test.go
@@ -18,7 +18,9 @@ func TestCreateClick(t *testing.T) {
 	clickService := clicks_service.NewClicksService(clickRepository)
 	mockRepository := mocks.NewMockUrlRepository()
 	mockService := url_service.NewURLService(mockRepository)
-	clickHandler := NewClickHandler(clickService, mockService)
+	tokenService := mocks.NewMockTokenService()
+
+	clickHandler := NewClickHandler(clickService, mockService, tokenService)
 
 	// Define test user data
 	t.Run("Should create click", func(t *testing.T) {
@@ -84,4 +86,146 @@ func TestCreateClick(t *testing.T) {
 		assert.NoError(t, err)
 	})
 
+}
+
+func TestGetUserClickDetails(t *testing.T) {
+	// Create mock user repository and service
+	clickRepository := mocks.NewMockClicksRepository()
+	clickService := clicks_service.NewClicksService(clickRepository)
+	mockRepository := mocks.NewMockUrlRepository()
+	mockService := url_service.NewURLService(mockRepository)
+	tokenService := mocks.NewMockTokenService()
+
+	clickHandler := NewClickHandler(clickService, mockService, tokenService)
+
+	t.Run("Success", func(t *testing.T) {
+		// Create a new Echo instance
+		e := echo.New()
+		req := httptest.NewRequest(http.MethodGet, "/:id/detail", nil)
+		rec := httptest.NewRecorder()
+		c := e.NewContext(req, rec)
+		c.SetPath("/:id")
+
+		req.Header.Set("Authorization", "Bearer valid")
+		c.SetParamNames("id")
+		c.SetParamValues("success")
+
+		// Mock GetUserClickDetails method
+		// Call GetUserClickDetailsHandler
+		err := clickHandler.GetUserClickDetailsHandler(c)
+
+		// Assertions
+		assert.Equal(t, http.StatusOK, rec.Code)
+		assert.Contains(t, rec.Body.String(), "[]")
+		assert.NoError(t, err)
+
+	})
+
+	t.Run("Should return error if token is not provided", func(t *testing.T) {
+		// Create a new Echo instance
+		e := echo.New()
+		req := httptest.NewRequest(http.MethodGet, "/:id/detail", nil)
+		rec := httptest.NewRecorder()
+		c := e.NewContext(req, rec)
+		c.SetPath("/:id")
+
+		c.SetParamNames("id")
+		c.SetParamValues("2")
+
+		// Mock GetUserClickDetails method
+		// Call GetUserClickDetailsHandler
+		err := clickHandler.GetUserClickDetailsHandler(c)
+
+		// Assertions
+		assert.Equal(t, http.StatusUnauthorized, rec.Code)
+		assert.Contains(t, rec.Body.String(), "error")
+		assert.NoError(t, err)
+	})
+
+	t.Run("Should return error if token is invalid", func(t *testing.T) {
+		// Create a new Echo instance
+		e := echo.New()
+		req := httptest.NewRequest(http.MethodGet, "/:id/detail", nil)
+		rec := httptest.NewRecorder()
+		c := e.NewContext(req, rec)
+		c.SetPath("/:id")
+
+		req.Header.Set("Authorization", "Bearer invalid")
+		c.SetParamNames("id")
+		c.SetParamValues("2")
+
+		// Mock GetUserClickDetails method
+		// Call GetUserClickDetailsHandler
+		err := clickHandler.GetUserClickDetailsHandler(c)
+
+		// Assertions
+		assert.Equal(t, http.StatusUnauthorized, rec.Code)
+		assert.Contains(t, rec.Body.String(), "error")
+		assert.NoError(t, err)
+	})
+
+	t.Run("Should return error if provided token other than Bearer", func(t *testing.T) {
+		// Create a new Echo instance
+		e := echo.New()
+		req := httptest.NewRequest(http.MethodGet, "/:id/detail", nil)
+		rec := httptest.NewRecorder()
+		c := e.NewContext(req, rec)
+		c.SetPath("/:id")
+
+		req.Header.Set("Authorization", "not invalid")
+		c.SetParamNames("id")
+		c.SetParamValues("2")
+
+		// Mock GetUserClickDetails method
+		// Call GetUserClickDetailsHandler
+		err := clickHandler.GetUserClickDetailsHandler(c)
+
+		// Assertions
+		assert.Equal(t, http.StatusUnauthorized, rec.Code)
+		assert.Contains(t, rec.Body.String(), "error")
+		assert.NoError(t, err)
+	})
+	t.Run("Should return error if id is invalid", func(t *testing.T) {
+		// Create a new Echo instance
+		e := echo.New()
+		req := httptest.NewRequest(http.MethodGet, "/:id/detail", nil)
+		rec := httptest.NewRecorder()
+		c := e.NewContext(req, rec)
+		c.SetPath("/:id")
+
+		req.Header.Set("Authorization", "Bearer valid")
+		c.SetParamNames("id")
+		c.SetParamValues("invalid")
+
+		// Mock GetUserClickDetails method
+		// Call GetUserClickDetailsHandler
+		err := clickHandler.GetUserClickDetailsHandler(c)
+
+		// Assertions
+		assert.Equal(t, http.StatusInternalServerError, rec.Code)
+		assert.Contains(t, rec.Body.String(), "error")
+		assert.NoError(t, err)
+	})
+
+	t.Run("Should return error if id is not valid", func(t *testing.T) {
+		// Create a new Echo instance
+		e := echo.New()
+		req := httptest.NewRequest(http.MethodGet, "/:id/detail", nil)
+		rec := httptest.NewRecorder()
+		c := e.NewContext(req, rec)
+		c.SetPath("/:id")
+
+		req.Header.Set("Authorization", "Bearer valid")
+		c.SetParamNames("id")
+		c.SetParamValues("not_valid")
+
+		// Mock GetUserClickDetails method
+		// Call GetUserClickDetailsHandler
+		err := clickHandler.GetUserClickDetailsHandler(c)
+
+		// Assertions
+		assert.Equal(t, http.StatusInternalServerError, rec.Code)
+		assert.Contains(t, rec.Body.String(), "error")
+		assert.NoError(t, err)
+	})
 }