+++ internal/app/repositories/auth/auth_repository.go
@@ -56,14 +56,14 @@ func (r *DBAuthRepository) Create(user *user_model.User) (*user_model.User, erro
 // GetByUsername retrieves an auth record from the database by username.
 func (r *DBAuthRepository) GetByUsername(username string) (*user_model.User, error) {
 	// Prepare SQL statement
-	query := "SELECT id, username, password, registration_date FROM users WHERE username = ?"
+	query := "SELECT id, username, password, created_at FROM users WHERE username = ?"
 	row := r.DB.QueryRow(query, username)
 
 	// Initialize a new User object to store the result
 	user := &user_model.User{}
 
 	// Scan the result into the User object
-	err := row.Scan(&user.ID, &user.Username, &user.Password, &user.RegistrationDate)
+	err := row.Scan(&user.ID, &user.Username, &user.Password, &user.CreatedAt)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
 			// Return a custom error if the auth with the specified username is not found