+++ internal/app/repositories/url/url_repository.go
@@ -3,6 +3,7 @@ package url_repository
 import (
 	"database/sql"
 	"errors"
+	"fmt"
 	"url-shortener/internal/app/models/url"
 )
 
@@ -10,6 +11,8 @@ import (
 type Repository interface {
 	CreateURL(originalURL, shortCode string, userId *uint) (string, error)
 	GetOriginalURL(shortCode string) (string, error)
+	GetUserURLs(userID uint) ([]url_model.URL, error)
+	GetUserWithShortURL(userID uint, shortURL string) error
 }
 
 // DBURLRepository is an implementation of URLRepository for MySQL database.
@@ -84,4 +87,53 @@ func (r *DBURLRepository) GetOriginalURL(shortCode string) (string, error) {
 	}
 
 	return originalURL, nil
+}
+
+// GetUserURLs retrieves the URLs created by the given user.
+func (r *DBURLRepository) GetUserURLs(userID uint) ([]url_model.URL, error) {
+	// Prepare SQL statement
+	query := "SELECT * FROM urls WHERE user_id = ?"
+	rows, err := r.DB.Query(query, userID)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+
+	// Initialize a slice to store the result
+	var urls []url_model.URL
+
+	// Iterate through the rows and scan the result into URL objects
+	for rows.Next() {
+		var u url_model.URL
+		err := rows.Scan(&u.OriginalURL, &u.ShortenedURL, &u.UserID, &u.CreatedAt)
+		if err != nil {
+			return nil, err
+		}
+		urls = append(urls, u)
+	}
+
+	return urls, nil
+}
+
+// GetUserWithShortURL retrieves the user who created the given shortened URL.
+func (r *DBURLRepository) GetUserWithShortURL(userID uint, shortURL string) error {
+	// Query to retrieve user_id associated with the short URL
+	query := "SELECT user_id FROM urls WHERE shortened_url = ?"
+	var retrievedUserID uint
+	err := r.DB.QueryRow(query, shortURL).Scan(&retrievedUserID)
+
+	// Handling errors
+	if err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			return url_model.ErrURLNotFound
+		}
+		return fmt.Errorf("error retrieving user ID: %v", err)
+	}
+
+	// Check if retrieved user_id matches the provided userID
+	if retrievedUserID != userID {
+		return fmt.Errorf("the provided short URL does not belong to the user")
+	}
+
+	return nil
 }