+++ internal/app/services/url/url_service_test.go
@@ -49,3 +49,51 @@ func TestGetOriginalURL(t *testing.T) {
 		assert.Error(t, err)
 	})
 }
+
+func TestGetUserUrls(t *testing.T) {
+	mockRepo := mocks.NewMockUrlRepository()
+
+	urlService := NewURLService(mockRepo)
+
+	t.Run("Get User URLs Successfully", func(t *testing.T) {
+		user := uint(1)
+		_, err := mockRepo.CreateURL("https://www.example.com", "abc123", &user)
+		if err != nil {
+			return
+		}
+		urls, err := urlService.GetUserURLs(1)
+		if err != nil {
+			t.Errorf("Error: %s", err)
+		}
+		assert.NotEmpty(t, urls)
+	})
+
+	t.Run("Should return error for nonexistent user", func(t *testing.T) {
+		_, err := urlService.GetUserURLs(2)
+		assert.Error(t, err)
+	})
+
+}
+
+func TestGetUserWithShortURL(t *testing.T) {
+	mockRepo := mocks.NewMockUrlRepository()
+
+	urlService := NewURLService(mockRepo)
+
+	t.Run("Get User with Short URL Successfully", func(t *testing.T) {
+		user := uint(1)
+		_, err := mockRepo.CreateURL("https://www.example.com", "abc123", &user)
+		if err != nil {
+			return
+		}
+		err = urlService.GetUserWithShortURL(1, "abc123")
+		if err != nil {
+			t.Errorf("Error: %s", err)
+		}
+	})
+
+	t.Run("Should return error for nonexistent URL", func(t *testing.T) {
+		err := urlService.GetUserWithShortURL(1, "invalid")
+		assert.Error(t, err)
+	})
+}