+++ internal/mocks/token.go
@@ -22,6 +22,12 @@ func (mts *MockTokenService) GenerateToken(user *user_model.User) (string, error
 	if user.Username == "error_token" {
 		return "", user_model.ErrUserAlreadyExists
 	}
+	if user.ID == 0 {
+		return "", url_model.ErrInvalidToken
+	}
+	if user.ID == 2 {
+		return "2", nil
+	}
 	// For simplicity in testing, return a fixed token
 	return "mockToken", nil
 }
@@ -30,6 +36,12 @@ func (mts *MockTokenService) GenerateToken(user *user_model.User) (string, error
 func (mts *MockTokenService) ValidateToken(tokenString string) (uint, error) {
 	if tokenString == "invalid" {
 		return 0, url_model.ErrInvalidToken
+	}
+	if tokenString == "expired" {
+		return 0, nil
+	}
+	if tokenString == "mockToken" {
+		return 1, nil
 	}
 	// For simplicity in testing, return a fixed user ID
 	return 123, nil