+++ internal/mocks/url.go
@@ -29,10 +29,17 @@ func (r *MockUrlRepository) CreateURL(originalUrl, shortCode string, userId *uin
 		}
 	}
 
+	var userID uint
+	if userId != nil {
+		userID = *userId
+	}
+
 	url := &url_model.URL{
 		OriginalURL:  originalUrl,
 		ShortenedURL: shortCode,
+		UserID:       userID,
 	}
+
 	r.Urls[uint(len(r.Urls)+1)] = url
 	return shortCode, nil
 }
@@ -53,4 +60,31 @@ func (r *MockUrlRepository) GetOriginalURL(shortCode string) (string, error) {
 	}
 	// Return an error if url not found
 	return "", url_model.ErrURLNotFound
+}
+
+// GetUserURLs simulates retrieving all urls created by a user from the mock database.
+func (r *MockUrlRepository) GetUserURLs(userId uint) ([]url_model.URL, error) {
+	urls := make([]url_model.URL, 0)
+	for _, u := range r.Urls {
+		if u.UserID == userId {
+			urls = append(urls, *u)
+		}
+	}
+	if len(urls) == 0 {
+		return nil, url_model.ErrURLNotFound
+	}
+	return urls, nil
+}
+
+// GetUserWithShortURL simulates checking if a user is the owner of an url in the mock database.
+func (r *MockUrlRepository) GetUserWithShortURL(userId uint, shortURL string) error {
+	if shortURL == "invalid" {
+		return url_model.ErrURLNotFound
+	}
+	for _, u := range r.Urls {
+		if u.UserID == userId && u.ShortenedURL == shortURL {
+			return nil
+		}
+	}
+	return nil
 }